+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to
 ## Changed
 - 💄(frontend) add filtering from left panel #475
 
+## Fixed
+- 🐛(frontend) update doc editor height #481
+
 
 ## [1.9.0] - 2024-12-11
 
@@ -48,9 +51,7 @@ and this project adheres to
 - 🐛(backend) fix sanitize problem IA #490
 
 
-
 ## [1.8.2] - 2024-11-28
-
 
 ## Changed